# server/api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: 이 문서는 스마트온실 UI를 위한 API 문서입니다.
  version: "0.3.0"
  title: Smart Greenhouse API
  contact:
    email: dev.jinong@gmail.com
host: "localhost:8081"
basePath: /common/v1
...
schemes:
  - http
consumes:
  - application/octet-stream
  - application/json
produces:
  - application/json  
securityDefinitions:
  Bearer:
    type: apiKey
    name: authorization
    in: header
paths:
  /mode:
    x-swagger-router-controller: fcore
    get:
      security:
        - Bearer: []
      tags:
        - fcore
        - p_intro
        - p_dashboard
        - p_basic
        - p_monitor
        - p_auto
        - p_manual
        - p_mypage
        - p_detail
      summary: 동작모드를 읽는다. 
      description: 동작모드를 확인한다. 모드가 manual 이거나 auto 가 될때 까지 기다려야 한다. 최대 30초가 소요될 수 있다.
      operationId: getMode
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/FarmosMode'
...
  /config:
    x-swagger-router-controller: fcore
    get:
      security:
        - Bearer: []
      tags:
        - fcore
        - p_intro
        - p_dashboard
        - p_basic
        - p_monitor
        - p_auto
        - p_manual
        - p_mypage
        - p_detail
      summary: 설정을 읽는다. 
      description: 모든 설정을 읽는다. file 인자가 주어지지 않으면 fcore.json 을 읽는다.
      operationId: getFarmosConfig
      produces:
        - application/json
      parameters:
        - in: query
          name: file
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/FarmosConfig'
    put:
      security:
        - Bearer: []
      tags:
        - fcore
        - p_detail
      summary: 설정을 수정한다.
      description: 'file 인자는 필수이다. 단, fcore.json 은 수정할 수 없다.'
      operationId: setFarmosConfig
      produces:
        - application/json
      parameters:
        - in: query
          name: file
          type: string
          required: true
        - in: body
          name: body
          description: configuration
          required: true
          schema:
            $ref: '#/definitions/FarmosConfig'
      responses:
        '200':
          description: successful operation
        '405':
          description: Invalid input
...
  '/field/{fieldId}/autocontrol':
    x-swagger-router-controller: fcore
    get:
      security:
        - Bearer: []
      tags:
        - field
        - p_intro
        - p_auto
        - p_manual
      summary: 자동제어와 관련된 설정을 읽는다. 
      description: 자동제어와 관련된 종합설정을 읽는다. 0번 구역에 대한 조회는 하지 않아야 한다.
      operationId: getFieldAutoControl
      produces:
        - application/json
      parameters:
        - name: fieldId
          in: path
          description: ID of field to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/FieldAutoControl'
        '405':
          description: Invalid input
    put:
      security:
        - Bearer: []
      tags:
        - field
        - p_intro
        - p_auto
        - p_manual
      summary: 자동제어와 관련된 설정을 수정한다.
      description: 자동제어와 관련된 설정을 수정한다. 0번 구역에 대한 수정은 하지 않아야 한다. mode는 auto, manual, local 중 하나를 선택한다. 
      operationId: setFieldAutoControl
      produces:
        - application/json
      parameters:
        - name: fieldId
          in: path
          description: ID of field to return
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: configuration
          required: true
          schema:
            $ref: '#/definitions/FieldAutoControl'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/FieldAutoControl'
        '405':
          description: Invalid input
...
  /farm/graph:
    x-swagger-router-controller: fcore
    put:
      security:
        - Bearer: []
      tags:
        - farm
        - p_dashboard
        - p_auto
        - p_basic
        - p_detail
        - p_monitor
      summary: 그래프를 그리기위한 정보를 획득
      description: 그래프 옵션으로 device_id만 입력하면 됨.
      operationId: getFarmGraph
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description : option for graph
          required: true
          schema:
            $ref: '#/definitions/GraphOption'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/GraphData'
        '405':
          description: Invalid input
  /farm/graph/download:
    x-swagger-router-controller: fcore
    put:
      security:
        - Bearer: []
      tags:
        - farm
        - p_monitor
      summary: 그래프 정보를 CSV로 다운로드 받음
      description: 그래프 옵션으로 device_id만 입력하면 됨.
      operationId: getFarmGraphDownload
      produces:
        - text/csv
      parameters:
        - name: body
          in: body
          description : option for graph
          required: true
          schema:
            $ref: '#/definitions/GraphOption'
      responses:
        '200':
          description: successful operation
          schema:
            type: file
        '405':
          description: Invalid input
  /farm/growth:
    x-swagger-router-controller: fcore
    get:
      security:
        - Bearer: []
      tags:
        - farm
      summary: 이미지를 얻어온다.
      description: ''
      operationId: getImage
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items: 
              type: string
        '405':
          description: Invalid input          
...
  '/field/{fieldId}/tcgraph':
    x-swagger-router-controller: fcore
    put:
      security:
        - Bearer: []
      tags:
        - field
        - p_auto
        - p_detail
      summary: 그래프를 그리기위한 TC 정보를 획득
      description: TC에 해당하는 device_id 를 보낸다. 동아이디 * 1000 + (524, 525, 526) 
      operationId: getFieldTCGraph
      produces:
        - application/json
      parameters:
        - name: fieldId
          in: path
          description: ID of field to return
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description : option for graph
          required: true
          schema:
            $ref: '#/definitions/GraphOption'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/TCGraphData'
        '405':
          description: Invalid input  
  '/field/{fieldId}/tcselect/{selecttc}':
    x-swagger-router-controller: fcore
    put:
      security:
        - Bearer: []
      tags:
        - field
        - p_auto
        - p_detail
      summary: TC 정보를 선택
      description: 'selecttc는 1, 2, 3 중에 하나여야 하며 각각 지난 1일간, 지난2일간, 지난3일간에 해당한다.'
      operationId: setFieldTC
      produces:
        - application/json
      parameters:
        - name: fieldId
          in: path
          description: ID of field to return
          required: true
          type: integer
          format: int64
        - name: selecttc
          in: path
          description : 온도보상 기준일을 선택
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
        '405':
          description: Invalid input
...
  '/user/login':
    x-swagger-router-controller: farmos
    put:
      tags:
        - user
        - p_mypage
      summary: 사용자 로그인
      description: userid 와 passwd만 채우면 된다. 다른 값들은 무시한다. 초기 일반사용자는 user/fcore 이고, 고급사용자는 admin/fcore이다. 일반사용자는 privilege가 user 이고, 고급사용자는 privilege가 admin이다. 
      operationId: loginUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: The password for login in clear text
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: successful operation
          schema:
             $ref: '#/definitions/User'
        '400':
          description: Invalid userid/password supplied
...
  '/user/logout/{userid}':
    x-swagger-router-controller: farmos
    get:
      security:
        - Bearer: []
      tags:
        - user
        - p_mypage
      summary: 사용자 로그 아웃
      description: 실제로 수행하는 로그아웃관련 행위는 없다. 그래서 아무값이나 넣어도 된다. :)
      operationId: logoutUser
      produces:
        - application/json
      parameters: 
        - name: userid
          in: path
          required: true
          type: string
      responses:
        default:
          description: successful operation
  '/user':
    x-swagger-router-controller: fcore
    put:
      security:
        - Bearer: []
      tags:
        - user
        - p_mypage
      summary: 사용자 정보 변경
      description: 암호를 변경하는경우는 newpasswd 라는 key를 사용하여 정보를 추가한다. userid와 passwd가 맞는경우 사용자 정보가 변경된다. 사용자권한 (privilege)는 변경할 수 없다.
      operationId: updateUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Updated user object
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '400':
          description: Invalid user supplied
        '404':
          description: User not found
...
definitions:
...
  GraphOption:
    type: object
    required : [device_id]
    properties:
      device_id:
        type: array
        items: 
          type: integer
        example: [111,211]
      startdate:
        type: string
        example: '2017-09-05 00:00:00'
      enddate:
        type: string
        example: '2017-09-05 23:00:00'
      extra:
        type: object
        properties:
          condensation:
            type: integer
            example: 2013
          CAI:
            type: string
            example: 'crop-activity-index-1.json'
  TCGraphData:
    type: object
    properties:
      selectedtc:
        type: integer
        example: 1
      data:
        type: array
        items:
          $ref: '#/definitions/GraphPointSet'
  GraphData:
    type: object
    required: [data]
    properties:
      data:
        type: array
        items:
          $ref: '#/definitions/GraphPointSet'
      extra:
        type: object
        properties:
          condensation:
            type: array
            items:
              $ref: '#/definitions/GraphPoint'
          CAI:
            type: array
            items:
              $ref: '#/definitions/GraphPoint'
  GraphPoint:
    type: object
    properties:
      time:
        type: string
        example: 1440
      value:
        type: number
        example: 32.1
  GraphPointSet:
    type: object
    properties:
      device_id:
        type: integer
        example: 112
      name:
        type: string
        example: 습도
      unit:
        type: string
        example: '%'
      data:
        type: array
        items:
          $ref: '#/definitions/GraphPoint'
    xml:
      name: GraphPointSet
...
  User:
    type: object
    properties:
      userid:
        type: string
        example: user
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      passwd:
        type: string
        example: fcore
      phone:
        type: string
      privilege:
        type: string
        description: User Status
        enum: 
          - admin
          - user
    xml:
      name: user
  FarmosConfig:
    type: object
    xml:
      name: config
  FarmosMode:
    type: object
    properties:
      mode:
        type: string
        enum: ["manual", "to manual", "to auto", "auto", "reload then manual", "reload then auto","initialize then manual", "initialize then auto"]
  FieldAutoControl:
    type: object
    properties:
      mode:
        type: string
        enum: [auto, manual, local]
      ventilationload:
        $ref: '#/definitions/AutoControlItem'
      heatload:
        $ref: '#/definitions/AutoControlItem'
      condensation:
        $ref: '#/definitions/AutoControlItem'
      badweather:
        $ref: '#/definitions/AutoControlItem'
      CAI:
        $ref: '#/definitions/AutoControlItem'
    xml:
      name: config
  AutoControlItem:
    type: object
    properties:
      use: 
        type: boolean
      option:
        type: object
...